+++ packages/toolbars/upload/meta.js
@@ -1,0 +1,11 @@
+export default {
+  id: 'engine.toolbars.upload',
+  type: 'toolbars',
+  title: 'upload',
+  options: {
+    icon: {
+      default: 'upload'
+    },
+    renderType: 'button'
+  }
+}
