+++ packages/toolbars/upload/src/styles/vars.less
@@ -1,0 +1,12 @@
+:root {
+  --te-toolbars-upload-button-bg-color: var(--te-common-bg-prompt);
+  --te-toolbars-upload-text-color-primary: var(--te-common-text-primary);
+  --te-toolbars-upload-text-color-secondary: var(--te-common-text-secondary);
+  --te-toolbars-upload-icon-color: var(--te-common-icon-secondary);
+  --te-toolbars-upload-icon-color-primary: var(--te-common-icon-primary);
+  --te-toolbars-upload-bg-color-primary: var(--te-common-bg-primary);
+  --te-toolbars-upload-bg-color: var(--te-common-bg-default);
+  --te-toolbars-upload-bg-color-hover: var(--te-common-bg-container);
+  --te-toolbars-upload-border-color-checked: var(--te-common-border-checked);
+  --te-toolbars-upload-border-color-divider: var(--te-common-border-divider);
+}
