+++ packages/vue-generator/test/testcases/generator/expected/appdemo01/src/views/testCanvasRowCol.vue
@@ -40,17 +40,8 @@ const { t, lowcodeWrap, stores } = vue.inject(I18nInjectionKey).lowcode()
 const wrap = lowcodeWrap(props, { emit })
 wrap({ stores })
 
-const state = vue.reactive({
-  canvasConfig: { rowGap: '20px', colGap: '20px', layoutMode: 'grid' },
-  uiState: { showGrid: true, selectedElement: null }
-})
+const state = vue.reactive({ dataDisk: [1, 2, 3] })
 wrap({ state })
-
-const handleElementClick = wrap(function handleElementClick(elementId) {
-  this.state.uiState.selectedElement = elementId
-})
-
-wrap({ handleElementClick })
 </script>
 <style scoped>
 body {
