+++ packages/vue-generator/test/testcases/generator/mockData.js
@@ -908,36 +908,7 @@ export const appSchemaDemo01 = {
     },
     {
       state: {
-        dataDisk: [1, 2, 3],
-        // 表单相关状态
-        formData: {
-          zone: '1',
-          cpu: '1',
-          memory: '1',
-          storageType: '1',
-          storageSize: '40',
-          diskType: '1',
-          diskSize: '100',
-          networkType: '1',
-          bandwidth: '1',
-          instanceType: '1',
-          instanceCount: '1'
-        },
-        // 输入框相关状态
-        inputValues: {
-          diskLabel: '',
-          systemDisk: '',
-          dataDiskSize: '',
-          networkConfig: ''
-        },
-        // 选择器相关状态
-        selectValues: {
-          availableZone: '1',
-          cpuArch: '1',
-          memorySize: '1',
-          storageOption: '1',
-          networkOption: '1'
-        }
+        dataDisk: [1, 2, 3]
       },
       methods: {},
       componentName: 'Page',
@@ -1084,11 +1055,7 @@ export const appSchemaDemo01 = {
                             value: '3'
                           }
                         ],
-                        modelValue: {
-                          type: 'JSExpression',
-                          model: true,
-                          value: 'state.selectValues.availableZone'
-                        }
+                        modelValue: '1'
                       },
                       id: '6184481b'
                     }
@@ -1137,11 +1104,7 @@ export const appSchemaDemo01 = {
                             value: '2'
                           }
                         ],
-                        modelValue: {
-                          type: 'JSExpression',
-                          model: true,
-                          value: 'state.formData.cpu'
-                        }
+                        modelValue: '1'
                       },
                       id: '7d33ced7'
                     }
@@ -1179,11 +1142,7 @@ export const appSchemaDemo01 = {
                             {
                               componentName: 'TinySelect',
                               props: {
-                                modelValue: {
-                                  type: 'JSExpression',
-                                  model: true,
-                                  value: 'state.selectValues.cpuArch'
-                                },
+                                modelValue: '',
                                 placeholder: '请选择',
                                 options: [
                                   {
@@ -1217,11 +1176,7 @@ export const appSchemaDemo01 = {
                             {
                               componentName: 'TinySelect',
                               props: {
-                                modelValue: {
-                                  type: 'JSExpression',
-                                  model: true,
-                                  value: 'state.selectValues.memorySize'
-                                },
+                                modelValue: '',
                                 placeholder: '请选择',
                                 options: [
                                   {
@@ -1249,19 +1204,15 @@ export const appSchemaDemo01 = {
                               componentName: 'Text',
                               props: {
                                 text: '规格名称',
-                                style: 'width: 120px;'
+                                style: 'width: 80px;'
                               },
                               id: 'd3eb6352'
                             },
                             {
                               componentName: 'TinySearch',
                               props: {
-                                placeholder: '输入关键词',
-                                modelValue: {
-                                  type: 'JSExpression',
-                                  model: true,
-                                  value: 'state.inputValues.diskLabel'
-                                }
+                                modelValue: '',
+                                placeholder: '输入关键词'
                               },
                               id: '21cb9282'
                             }
@@ -1466,11 +1417,7 @@ export const appSchemaDemo01 = {
                         {
                           componentName: 'TinySelect',
                           props: {
-                            modelValue: {
-                              type: 'JSExpression',
-                              model: true,
-                              value: 'state.selectValues.storageOption'
-                            },
+                            modelValue: '',
                             placeholder: '请选择',
                             options: [
                               {
@@ -1489,11 +1436,7 @@ export const appSchemaDemo01 = {
                         {
                           componentName: 'TinySelect',
                           props: {
-                            modelValue: {
-                              type: 'JSExpression',
-                              model: true,
-                              value: 'state.selectValues.networkOption'
-                            },
+                            modelValue: '',
                             placeholder: '请选择',
                             options: [
                               {
@@ -1572,11 +1515,7 @@ export const appSchemaDemo01 = {
                         {
                           componentName: 'TinySelect',
                           props: {
-                            modelValue: {
-                              type: 'JSExpression',
-                              model: true,
-                              value: 'state.formData.storageType'
-                            },
+                            modelValue: '',
                             placeholder: '请选择',
                             options: [
                               {
@@ -1596,11 +1535,7 @@ export const appSchemaDemo01 = {
                           componentName: 'TinyInput',
                           props: {
                             placeholder: '请输入',
-                            modelValue: {
-                              type: 'JSExpression',
-                              model: true,
-                              value: 'state.inputValues.systemDisk'
-                            },
+                            modelValue: '',
                             style: 'width: 120px; margin-right: 10px;'
                           },
                           id: '1cde4c0f'
@@ -1657,11 +1592,7 @@ export const appSchemaDemo01 = {
                         {
                           componentName: 'TinySelect',
                           props: {
-                            modelValue: {
-                              type: 'JSExpression',
-                              model: true,
-                              value: 'state.formData.diskType'
-                            },
+                            modelValue: '',
                             placeholder: '请选择',
                             options: [
                               {
@@ -1681,11 +1612,7 @@ export const appSchemaDemo01 = {
                           componentName: 'TinyInput',
                           props: {
                             placeholder: '请输入',
-                            modelValue: {
-                              type: 'JSExpression',
-                              model: true,
-                              value: 'state.inputValues.dataDiskSize'
-                            },
+                            modelValue: '',
                             style: 'width: 120px; margin-right: 10px;'
                           },
                           id: '667c7926'
@@ -1702,11 +1629,7 @@ export const appSchemaDemo01 = {
                           componentName: 'TinyInput',
                           props: {
                             placeholder: '请输入',
-                            modelValue: {
-                              type: 'JSExpression',
-                              model: true,
-                              value: 'state.inputValues.diskLabel'
-                            },
+                            modelValue: '',
                             style: 'width: 120px;'
                           },
                           id: '1bd56dc0'
@@ -1818,11 +1741,7 @@ export const appSchemaDemo01 = {
                               componentName: 'TinyInput',
                               props: {
                                 placeholder: '请输入',
-                                modelValue: {
-                                  type: 'JSExpression',
-                                  model: true,
-                                  value: 'state.formData.instanceCount'
-                                },
+                                modelValue: '',
                                 style: 'width: 120px; margin-right: 10px;'
                               },
                               id: '2f9cf442'
@@ -2108,23 +2027,6 @@ export const appSchemaDemo01 = {
         }
       ],
       fileName: 'testCanvasRowCol',
-      state: {
-        canvasConfig: {
-          rowGap: '20px',
-          colGap: '20px',
-          layoutMode: 'grid'
-        },
-        uiState: {
-          showGrid: true,
-          selectedElement: null
-        }
-      },
-      methods: {
-        handleElementClick: {
-          type: 'JSFunction',
-          value: 'function handleElementClick(elementId) { this.state.uiState.selectedElement = elementId; }'
-        }
-      },
       meta: {
         name: 'testCanvasRowCol',
         id: 1737797330916,
@@ -2187,142 +2089,6 @@ export const appSchemaDemo01 = {
         router: 'createVm/untitledFA/UntitledA'
       },
       path: 'createVm/untitledFA'
-    },
-    {
-      componentName: 'div',
-      fileName: 'LifeCyclePage',
-      meta: {
-        id: 'lifecycle-page',
-        isPage: true,
-        parentId: '0',
-        router: '/lifecycle'
-      },
-      props: {
-        class: 'lifecycle-container'
-      },
-      children: [
-        {
-          componentName: 'TinyCard',
-          props: {
-            title: '仪表板'
-          },
-          children: [
-            {
-              componentName: 'div',
-              props: {
-                class: 'metrics-grid'
-              },
-              children: [
-                {
-                  componentName: 'div',
-                  props: {
-                    class: 'metric-item'
-                  },
-                  children: [
-                    {
-                      componentName: 'Text',
-                      props: {
-                        text: {
-                          type: 'JSExpression',
-                          value: '"活跃用户: " + this.state.activeUsers'
-                        }
-                      }
-                    }
-                  ]
-                },
-                {
-                  componentName: 'div',
-                  props: {
-                    class: 'metric-item'
-                  },
-                  children: [
-                    {
-                      componentName: 'Text',
-                      props: {
-                        text: {
-                          type: 'JSExpression',
-                          value: '"系统负载: " + this.state.systemLoad + "%"'
-                        }
-                      }
-                    }
-                  ]
-                }
-              ]
-            },
-            {
-              componentName: 'TinyButton',
-              props: {
-                type: 'success',
-                onClick: {
-                  type: 'JSExpression',
-                  value: 'this.refreshData'
-                }
-              },
-              children: ['刷新数据']
-            }
-          ]
-        }
-      ],
-      state: {
-        activeUsers: 0,
-        systemLoad: 0,
-        lastUpdateTime: null,
-        autoRefresh: true
-      },
-      methods: {
-        refreshData: {
-          type: 'JSFunction',
-          value: 'function refreshData() { this.fetchMetrics(); this.state.lastUpdateTime = new Date(); }'
-        },
-        fetchMetrics: {
-          type: 'JSFunction',
-          value:
-            'function fetchMetrics() { this.state.activeUsers = Math.floor(Math.random() * 1000); this.state.systemLoad = Math.floor(Math.random() * 100); }'
-        },
-        startAutoRefresh: {
-          type: 'JSFunction',
-          value:
-            'function startAutoRefresh() { if (this.autoRefresh) { this.refreshInterval = setInterval(() => this.refreshData(), 30000); } }'
-        },
-        stopAutoRefresh: {
-          type: 'JSFunction',
-          value:
-            'function stopAutoRefresh() { if (this.refreshInterval) { clearInterval(this.refreshInterval); this.refreshInterval = null; } }'
-        },
-        handleVisibilityChange: {
-          type: 'JSFunction',
-          value:
-            'function handleVisibilityChange() { if (document.hidden) { this.stopAutoRefresh(); } else { this.startAutoRefresh(); } }'
-        }
-      },
-      lifeCycles: {
-        onMounted: {
-          type: 'JSFunction',
-          value:
-            'function onMounted() { console.log("onMounted."); refreshData(); startAutoRefresh(); document.addEventListener("visibilitychange", handleVisibilityChange) }'
-        },
-        onUnmounted: {
-          type: 'JSFunction',
-          value:
-            'function onUnmounted() { console.log("onUnmounted."); stopAutoRefresh(); document.removeEventListener("visibilitychange", handleVisibilityChange) }'
-        },
-        onActivated: {
-          type: 'JSFunction',
-          value: 'function onActivated() { console.log("页面激活"); refreshData() }'
-        },
-        onDeactivated: {
-          type: 'JSFunction',
-          value: 'function onDeactivated() { console.log("页面停用"); stopAutoRefresh() }'
-        },
-        onUpdated: {
-          type: 'JSFunction',
-          value: 'function onUpdated() { console.log("页面已更新", this.state.loginCount) }'
-        },
-        onBeforeMount: {
-          type: 'JSFunction',
-          value: 'function onBeforeMount() { console.log("onBeforeMount.") }'
-        }
-      }
     }
   ],
   componentsMap: [
@@ -2331,28 +2097,28 @@ export const appSchemaDemo01 = {
       package: '@opentiny/vue',
       exportName: 'CarouselItem',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyCheckboxButton',
       package: '@opentiny/vue',
       exportName: 'CheckboxButton',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.17'
     },
     {
       componentName: 'TinyTree',
       package: '@opentiny/vue',
       exportName: 'Tree',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyPopover',
       package: '@opentiny/vue',
       exportName: 'Popover',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyTooltip',
@@ -2366,21 +2132,21 @@ export const appSchemaDemo01 = {
       package: '@opentiny/vue',
       exportName: 'Col',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyDropdownItem',
       package: '@opentiny/vue',
       exportName: 'DropdownItem',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyPager',
       package: '@opentiny/vue',
       exportName: 'Pager',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyPlusAccessdeclined',
@@ -2415,21 +2181,21 @@ export const appSchemaDemo01 = {
       package: '@opentiny/vue',
       exportName: 'Search',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.13'
     },
     {
       componentName: 'TinyRow',
       package: '@opentiny/vue',
       exportName: 'Row',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyFormItem',
       package: '@opentiny/vue',
       exportName: 'FormItem',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyAlert',
@@ -2443,21 +2209,21 @@ export const appSchemaDemo01 = {
       package: '@opentiny/vue',
       exportName: 'Input',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyTabs',
       package: '@opentiny/vue',
       exportName: 'Tabs',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyDropdownMenu',
       package: '@opentiny/vue',
       exportName: 'DropdownMenu',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyDialogBox',
@@ -2471,98 +2237,91 @@ export const appSchemaDemo01 = {
       package: '@opentiny/vue',
       exportName: 'Switch',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyTimeLine',
       package: '@opentiny/vue',
       exportName: 'TimeLine',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyTabItem',
       package: '@opentiny/vue',
       exportName: 'TabItem',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyRadio',
       package: '@opentiny/vue',
       exportName: 'Radio',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyForm',
       package: '@opentiny/vue',
       exportName: 'Form',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyGrid',
       package: '@opentiny/vue',
       exportName: 'Grid',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyNumeric',
       package: '@opentiny/vue',
       exportName: 'Numeric',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyCheckboxGroup',
       package: '@opentiny/vue',
       exportName: 'CheckboxGroup',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.17'
     },
     {
       componentName: 'TinySelect',
       package: '@opentiny/vue',
       exportName: 'Select',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyButtonGroup',
       package: '@opentiny/vue',
       exportName: 'ButtonGroup',
       destructuring: true,
-      version: '3.24.0'
-    },
-    {
-      componentName: 'TinyButton',
-      package: '@opentiny/vue',
-      exportName: 'Button',
-      destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyCarousel',
       package: '@opentiny/vue',
       exportName: 'Carousel',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyPopeditor',
       package: '@opentiny/vue',
       exportName: 'Popeditor',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyDatePicker',
       package: '@opentiny/vue',
       exportName: 'DatePicker',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'TinyDropdown',
@@ -2576,7 +2335,7 @@ export const appSchemaDemo01 = {
       package: '@opentiny/vue',
       exportName: 'ChartHistogram',
       destructuring: true,
-      version: '3.24.0'
+      version: '0.1.16'
     },
     {
       componentName: 'PortalHome',
@@ -2607,13 +2366,6 @@ export const appSchemaDemo01 = {
       main: '',
       destructuring: false,
       version: '1.0.0'
-    },
-    {
-      componentName: 'TinyCard',
-      exportName: 'Card',
-      package: '@opentiny/vue',
-      version: '^3.10.0',
-      destructuring: true
     }
   ],
   meta: {
